:recycle:[refactor] : product 페이지 강제 라우팅
